requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
google-cloud-storage==2.9.0
PyMySQL
